requirements.txt
django==5.0
whitenoise 
mysqlclient
django_momoapi
qrcode
requests
opencv-python 
pyzbar
qrcode[pil]  
Pillow
django-axes
twilio
django-twilio
python-decouple
choco install ngrok
uni-sdk
django-ratelimit
django-redis
unimatrix
firebase-admin

